Add config package
