+++ internal/api/config/flags.go
@@ -1,4 +1,4 @@
-package api
+package config
 
 import (
 	"github.com/urfave/cli/v2"
